requirements.txt
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
pydantic==2.5.0
pydantic-settings==2.1.0